Script fixes, Init UE Tau Fix
