+++ Documentation/Documentation.md
@@ -238,6 +238,20 @@ f.  mmeSmDefs.cpp - Populate the eventIdToString and StateIdToString Maps, which
   *  For Integer Types, it is assigned as 0 implicitly.  
   *  For Array Types, the members are memset in the constructor.  
   *  For Pointer Types, it is assigned as NULL.  
+  
+  e) "Properties" :
+
+  * It's an array where you can specify various properties for the data member. Possible parameters are listed below:
+	  * metrics : "counter"/"gague".
+	    * If metrics value is "counter", the script will generate a method to increment the data member.
+	    * If metrics value is "guage", the script will generate both increment and decrement method for data member.
+    
+		  <pre>eg :
+			 {
+				 "Name": "dwnLnkSeqNo",
+				 "Type": "uint16_t",
+				 "Properties" : [ {"metrics":"counter"}]
+			 }</pre>
 
 2.  Specify the components to be generated by the codeGen in the generationItem.json in below format:
