+++ src/mme-app/actionHandlers/networkInitDetachActionHandlers.cpp
@@ -50,7 +50,7 @@ ActStatus ActionHandlers::ni_detach_req_to_ue(SM::ControlBlock& cb)
 	ni_detach_req.enb_s1ap_ue_id =  ue_ctxt->getS1apEnbUeId();
 	ni_detach_req.detach_type = 00000010;
 	
-	ue_ctxt->setDwnLnkSeqNo(ue_ctxt->getDwnLnkSeqNo()+1);
+	ue_ctxt->incrementDwnLnkSeqNo();
 	ni_detach_req.dl_seq_no = ue_ctxt->getDwnLnkSeqNo();
 	
 	memcpy(&(ni_detach_req.int_key), &(ue_ctxt->getUeSecInfo().secinfo_m.int_key), NAS_INT_KEY_SIZE);
@@ -80,7 +80,7 @@ ActStatus ActionHandlers::process_detach_accept_from_ue(SM::ControlBlock& cb)
 		return ActStatus::HALT;
 	}
 		
-	ue_ctxt->setUpLnkSeqNo(ue_ctxt->getUpLnkSeqNo()+1);
+	ue_ctxt->incrementUpLnkSeqNo();
 	
 	log_msg(LOG_DEBUG, "Leaving process_detach_accept_from_ue \n");
 	ProcedureStats::num_of_detach_accept_from_ue ++;
