+++ src/mme-app/actionHandlers/tauActionHandlers.cpp
@@ -42,60 +42,6 @@ using namespace cmn;
 using namespace cmn::utils;
 
 /***************************************
-* Action handler : process_tau_request
-***************************************/
-ActStatus ActionHandlers::process_tau_request(ControlBlock& cb)
-{
-	log_msg(LOG_INFO,"Inside process_tau_request\n");
-
-	UEContext *ue_ctxt = static_cast<UEContext*>(cb.getPermDataBlock());
-	if (ue_ctxt == NULL)
-	{
-		log_msg(LOG_ERROR, "process_tau_request: ue context is NULL\n",cb.getCBIndex());
-		return ActStatus::HALT;
-	}
-
-	MmeProcedureCtxt* prcdCtxt_p = dynamic_cast<MmeProcedureCtxt*>(cb.getTempDataBlock());
-	if (prcdCtxt_p == NULL)
-	{
-		log_msg(LOG_DEBUG, "process_tau_request: MmeProcedureCtxt is NULL\n");
-		return ActStatus::HALT;
-	}
-
-	MsgBuffer* msgBuf = static_cast<MsgBuffer*>(cb.getMsgData());
-	if (msgBuf == NULL)
-	{
-            log_msg(LOG_DEBUG,"process_tau_req: msgBuf is NULL \n");
-            return ActStatus::HALT;
-	}
-
-	const s1_incoming_msg_data_t* msgData_p =
-			static_cast<const s1_incoming_msg_data_t*>(msgBuf->getDataPointer());
-	if (msgData_p == NULL)
-	{
-		log_msg(LOG_ERROR, "Failed to retrieve data buffer \n");
-		return ActStatus::HALT;
-	}
-
-	const struct tauReq_Q_msg &tauReq = (msgData_p->msg_data.tauReq_Q_msg_m);
-	ue_ctxt->setUpLnkSeqNo(ue_ctxt->getUpLnkSeqNo()+1);
-
-	if( prcdCtxt_p->getCtxtType() == s1Handover_c)
-	{
-		S1HandoverProcedureContext *s1HoPrCtxt = dynamic_cast<S1HandoverProcedureContext*>(prcdCtxt_p);
-
-		//TAI and CGI obtained from s1ap ies.
-		//Convert the plmn in s1ap format to nas format
-		//before storing in ue context/sending in tai list of tau response.
-		MmeCommonUtils::formatS1apPlmnId(const_cast<PLMN*>(&tauReq.tai.plmn_id));
-		MmeCommonUtils::formatS1apPlmnId(const_cast<PLMN*>(&tauReq.eUtran_cgi.plmn_id));
-		s1HoPrCtxt->setTargetTai(Tai(tauReq.tai));
-		s1HoPrCtxt->setTargetCgi(Cgi(tauReq.eUtran_cgi));
-	}
-    return ActStatus::PROCEED;
-}
-
-/***************************************
 * Action handler : send_tau_response_to_ue
 ***************************************/
 ActStatus ActionHandlers::send_tau_response_to_ue(ControlBlock& cb)
@@ -138,7 +84,7 @@ ActStatus ActionHandlers::send_tau_response_to_ue(ControlBlock& cb)
 	tau_resp.status = 0;
 	tau_resp.ue_idx = ue_ctxt->getContextID();
 	tau_resp.dl_seq_no = ue_ctxt->getDwnLnkSeqNo();
-	ue_ctxt->setDwnLnkSeqNo(tau_resp.dl_seq_no+1);
+	ue_ctxt->incrementDwnLnkSeqNo();
 	memcpy(&(tau_resp.int_key), &(ue_ctxt->getUeSecInfo().secinfo_m.int_key),
 			NAS_INT_KEY_SIZE);
