+++ src/mme-app/contextManager/dataBlocks.cpp
@@ -216,6 +216,14 @@ namespace mme
 	{
 		return dwnLnkSeqNo_m;
 	}
+    
+	/******************************************************************************
+	* increment dwnLnkSeqNo
+	******************************************************************************/	
+	void UEContext::incrementDwnLnkSeqNo()
+	{
+		dwnLnkSeqNo_m = dwnLnkSeqNo_m + 1;
+	}   
 	
 	/******************************************************************************
 	* sets upLnkSeqNo
@@ -232,6 +240,14 @@ namespace mme
 	{
 		return upLnkSeqNo_m;
 	}
+    
+	/******************************************************************************
+	* increment upLnkSeqNo
+	******************************************************************************/	
+	void UEContext::incrementUpLnkSeqNo()
+	{
+		upLnkSeqNo_m = upLnkSeqNo_m + 1;
+	}   
 	
 	/******************************************************************************
 	* sets ueState
@@ -932,6 +948,7 @@ namespace mme
 	{
 		return pcoOptions_m;
 	}
+
 	/******************************************************************************
 	* returns pcoOptionsLen
 	******************************************************************************/
